fix: #66 make sidebars active when modal is open
